modified data type.

+++ libah/os.h
@@ -3,7 +3,7 @@
 #include "condition.h"
 #include "stdint.h"
 
-__host void    _h_sleepmilli(int64_t milli);
-__host int64_t _h_nowmilli();
+__host void _h_sleepmilli(long long milli);
+__host long long _h_nowmilli();
 
 __host int _h_system(const char *cmd);
